pyproject.toml
[project]
name = "hihiton"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "deepagents>=0.2.5",
    "fuzzywuzzy>=0.18.0",
    "langchain>=1.0.8",
    "langchain-community>=0.4.1",
    "langchain-mcp-adapters>=0.1.13",
    "langchain-openai>=1.0.3",
    "langfuse>=3.10.1",
    "langgraph>=1.0.3",
    "loguru>=0.7.3",
    "mcp>=1.22.0",
    "psycopg2-binary>=2.9.11",
    "pydantic>=2.12.4",
    "python-dotenv>=1.2.1",
    "httpx>=0.28.1",
    "beautifulsoup4>=4.14.2",
]
